// frontend/src/generated/types/dateTime.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface DateTime {
  NanoSeconds?: number;
  Seconds?: number;
  Minutes?: number;
  Hours?: number;
  Day?: number;
  Month?: number;
  Year?: number;
  IsUTC?: boolean;
}